feat: refactor RoomStage handling to use value objects for improved clarity
