+++ src/main/java/com/icebreaker/be/application/room/RoomStageService.java
@@ -1,8 +1,7 @@
 package com.icebreaker.be.application.room;
 
-import com.icebreaker.be.domain.room.entity.RoomStage;
-import com.icebreaker.be.domain.room.entity.Stage;
-import com.icebreaker.be.domain.room.repo.RoomStageRepository;
+import com.icebreaker.be.application.room.statemachine.RoomStageStateMachine;
+import com.icebreaker.be.domain.room.vo.StageEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,13 +9,9 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class RoomStageService {
 
-    private final RoomStageRepository roomStageRepository;
+    private final RoomStageStateMachine roomStageStateMachine;
 
-    /**
-     * 이벤트를 통해서만 접근 가능 RoomCode가 없을 수 없음 (앞에서 기저 처리)
-     */
-    public void changeStage(String roomCode, Stage stage) {
-        RoomStage roomStage = RoomStage.of(roomCode, stage);
-        roomStageRepository.save(roomStage);
+    public void changeStage(String roomCode, StageEvent event) {
+        roomStageStateMachine.transitionStage(roomCode, event);
     }
 }
